feat: useMonaco hook
